cs2-arbitrage/requirements.txt
aiohttp>=3.9.0
websockets>=12.0
python-socketio[asyncio_client]>=5.11.0
httpx>=0.27.0
rich>=13.7.0
python-dotenv>=1.0.0
pandas>=2.2.0
orjson>=3.9.0
